Add files via upload
Sliders that have a bounding box larger than 2^14 in either dimension (and are therefore distorted) are no longer affected by the invisibility process. This means that repeated application of the tool to a single map will not result in distorted sliders slowly being pushed further and further back in time.